+++ CHANGELOG.md
@@ -4,13 +4,30 @@
 
 [1]: https://pypi.org/project/demisto-sdk/#history
 
+### 0.4.0
+* Added a validator for Indicator Types (reputation-*.json).
+* Added a BC validation for changing incident field type.
+* Fixed an issue where init command would generate an invalid yml for scripts.
+* Fixed an issue in misleading error message in v2 validation hook.
+* Fixed an issue in v2 hook which now is set only on newly added scripts.
+
+### 0.3.10
+* Added a BC check for incident fields - changing from version is not allowed.
+* Fixed an issue in create-content-artifacts where scripts in Packs in TestPlaybooks dir were copied with a wrong prefix.
+
+
 ### 0.3.9
+* Added a validation that incident field can not be required.
+* Added validation for fetch incident parameters.
+* Added validation for feed integration parameters.
+* Added to the *format* command the deletion of the *sourceplaybookid* field.
+* Fixed an issue where *fieldMapping* in playbook did not pass the scheme validation.
 * Fixed an issue where *create-content-artifacts* did not copy TestPlaybooks in Packs without prefix of *playbook-*.
 * Added a validation the a playbook can not have a rolename set.
 * Added to the image validator the new DBot default image.
 * Added the fields: elasticcommonfields, quiet, quietmode to the Playbook schema.
 * Fixed an issue where *validate* failed on integration commands without outputs.
-
+* Added a new hook for naming of v2 integrations and scripts.
 
 ### 0.3.8
 * Fixed an issue where *create-content-artifact* was not loading the data in the yml correctly.